📝 📦 Chore, Struct: 项目架构调整
